initial rework...
